+++ evaluator/src/lib.rs
@@ -14,6 +14,12 @@ pub mod normalizer;
 pub mod picker;
 pub mod progress;
 pub mod rand;
+pub mod repl;
 pub mod simulator;
 pub mod storage;
+pub mod tester;
+pub mod trace_quality;
 pub mod value;
+mod verbosity;
+
+pub use verbosity::Verbosity;